ДЗ к Уроку 6

+++ Controllers/PersonsController.cs
@@ -1,6 +1,10 @@
 ﻿using System.Threading.Tasks;
 using Lesson3.Controllers.Models;
+using Lesson3.Requests;
+using Lesson3.Requests.PersonRequests;
 using Lesson3.Services;
+using Lesson3.Validation;
+using Lesson3.Validation.Requests;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Lesson3.Controllers
@@ -11,16 +15,37 @@ namespace Lesson3.Controllers
     public class PersonsController : ControllerBase
     {
         private readonly IService<PersonDto> _service;
+        private readonly IPersonByIdValidator _personByIdValidator;
+        private readonly IPersonByNameValidator _personByNameValidator;
+        private readonly IPersonValidator _personValidator;
+        private readonly ISelectValidator _selectValidator;
 
-        public PersonsController(IService<PersonDto> service)
+        public PersonsController(IService<PersonDto> service,
+            IPersonByIdValidator personByIdValidator,
+            IPersonByNameValidator personByNameValidator,
+            IPersonValidator personValidator,
+            ISelectValidator selectValidator)
         {
             _service = service;
+            _personByIdValidator = personByIdValidator;
+            _personByNameValidator = personByNameValidator;
+            _personValidator = personValidator;
+            _selectValidator = selectValidator;
         }
 
         // GET /persons/{id} — получение человека по идентификатору
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            var request = new PersonByIdRequest(){Id = id};
+            
+            var validation = new OperationResult<PersonByIdRequest>(_personByIdValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             return Ok(await _service.Get(id));
         }
 
@@ -28,6 +53,12 @@ namespace Lesson3.Controllers
         [HttpGet("search")]
         public async Task<IActionResult> GetByName([FromQuery] string searchTerm)
         {
+            var request = new PersonByNameRequest() {Name = searchTerm};
+            var validation = new OperationResult<PersonByNameRequest>(_personByNameValidator.ValidateEntity(request));
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
             return Ok(await _service.GetByName(searchTerm));
         }
 
@@ -35,6 +66,14 @@ namespace Lesson3.Controllers
         [HttpGet]
         public async Task<IActionResult> Select([FromQuery] int skip, [FromQuery] int take)
         {
+            var request = new SelectionRequest() {Skip = skip, Take = take};
+
+            var validation = new OperationResult<SelectionRequest>(_selectValidator.ValidateEntity(request));
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+
+            }
             return Ok(await _service.Select(skip, take));
         }
 
@@ -43,6 +82,11 @@ namespace Lesson3.Controllers
         [HttpPost]
         public async Task<IActionResult> AddPerson([FromBody]PersonDto person)
         {
+            var validation = new OperationResult<PersonDto>(_personValidator.ValidateEntity(person));
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
             await _service.Create(person);
             return Ok();
         }
@@ -52,6 +96,11 @@ namespace Lesson3.Controllers
         [HttpPut]
         public async Task<IActionResult> UpdatePerson([FromBody] PersonDto person)
         {
+            var validation = new OperationResult<PersonDto>(_personValidator.ValidateEntity(person));
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
             await _service.Update(person);
             return Ok();
         }
@@ -61,6 +110,15 @@ namespace Lesson3.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeletePerson([FromRoute] int id)
         {
+            var request = new PersonByIdRequest(){Id = id};
+            
+            var validation = new OperationResult<PersonByIdRequest>(_personByIdValidator.ValidateEntity(request));
+
+            if (!validation.Succeed)
+            {
+                return BadRequest(validation);
+            }
+
             await _service.Delete(id);
             return Ok();
         }
